Introduced FrameCodec and moves IOBuffer frame constants.
