+++ hazelcast/src/main/java/com/hazelcast/bulktransport/impl/InitBulkTransportOp.java
@@ -5,7 +5,7 @@ import com.hazelcast.tpc.requestservice.Op;
 import com.hazelcast.tpc.requestservice.OpCodes;
 import com.hazelcast.table.impl.TableManager;
 
-import static com.hazelcast.tpc.engine.iobuffer.IOBuffer.OFFSET_REQ_CALL_ID;
+import static com.hazelcast.tpc.requestservice.FrameCodec.OFFSET_REQ_CALL_ID;
 
 public class InitBulkTransportOp extends Op {
