+++ hazelcast/src/main/java/com/hazelcast/tpc/requestservice/RequestService.java
@@ -58,7 +58,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
 
 import static com.hazelcast.internal.util.HashUtil.hashToIndex;
-import static com.hazelcast.tpc.engine.iobuffer.IOBuffer.OFFSET_REQ_CALL_ID;
+import static com.hazelcast.tpc.requestservice.FrameCodec.OFFSET_REQ_CALL_ID;
 import static java.util.concurrent.TimeUnit.SECONDS;
